Moved main.go and cmd folder to root of the repo
