+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/pegnet/pegnet/pegnet/cmd"
+import "github.com/pegnet/pegnet/cmd"
 
 func main() {
 	cmd.Execute()
